Use `name` instead of `property` attribute in meta tag
Lighthouse (and probably other tools) do not detect `<meta name="description" ... >` but only `<meta property="description" ... >`. According to [this SO answer](https://stackoverflow.com/a/22418095/7196903), the latter is the standard HTML5 way to provide a site's description metadata, while the former is the RDFa way to do the same. According to [another SO answer](https://stackoverflow.com/a/20429551/7196903), it is absolutely valid to provide both attributes in the same meta tag, so here we go.